+++ libraries/muxer/src/test/java/androidx/media3/muxer/Mp4MuxerEndToEndTest.java
@@ -26,6 +26,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.util.Pair;
 import androidx.media3.common.C;
@@ -1008,6 +1009,59 @@ public class Mp4MuxerEndToEndTest {
         new DumpableMp4Box(ByteBuffer.wrap(TestUtil.getByteArrayFromFilePath(outputFilePath)));
     DumpFileAsserts.assertOutput(
         context, dumpableBox, MuxerTestUtil.getExpectedDumpFilePath("mp4_with_t35_metadata.box"));
+  }
+
+  @Test
+  public void writeMp4File_withT35MetadataTrackReferencingVideoTrack_matchesExpectedBoxStructure()
+      throws Exception {
+    String outputFilePath = temporaryFolder.newFile().getPath();
+    // ITU-T T.35 prefix: country code, provider code, etc.
+    byte[] t35Prefix = new byte[] {0x00, 0x01, 0x02, 0x03, 0x04};
+    Format t35Format =
+        new Format.Builder()
+            .setSampleMimeType(MimeTypes.APPLICATION_ITUT_T35)
+            .setInitializationData(ImmutableList.of(t35Prefix))
+            .build();
+
+    try (Mp4Muxer muxer = new Mp4Muxer.Builder(SeekableMuxerOutput.of(outputFilePath)).build()) {
+      muxer.addMetadataEntry(
+          new Mp4TimestampData(
+              /* creationTimestampSeconds= */ 1_000_000L,
+              /* modificationTimestampSeconds= */ 5_000_000L));
+      int videoTrackId = muxer.addTrack(FAKE_VIDEO_FORMAT);
+      writeFakeSamples(muxer, videoTrackId, /* sampleCount= */ 5);
+      int audioTrackId = muxer.addTrack(FAKE_AUDIO_FORMAT);
+      writeFakeSamples(muxer, audioTrackId, /* sampleCount= */ 5);
+      int t35TrackId = muxer.addTrack(t35Format);
+      writeFakeSamples(muxer, t35TrackId, /* sampleCount= */ 5);
+      muxer.addTrackReference(
+          t35TrackId, Mp4Muxer.TRACK_REFERENCE_TYPE_CDSC, ImmutableList.of(videoTrackId));
+    }
+
+    DumpableMp4Box dumpableBox =
+        new DumpableMp4Box(ByteBuffer.wrap(TestUtil.getByteArrayFromFilePath(outputFilePath)));
+    // Last `trak` box (corresponding to T35 metadata track) contains `tref` box.
+    DumpFileAsserts.assertOutput(
+        context,
+        dumpableBox,
+        MuxerTestUtil.getExpectedDumpFilePath(
+            "mp4_with_t35_metadata_track_referencing_video_track.box"));
+  }
+
+  @Test
+  @SuppressLint("WrongConstant") // Intentionally assigned wrong constant.
+  public void addTrackReference_withUnsupportedTrackReferenceType_throws() throws Exception {
+    String outputFilePath = temporaryFolder.newFile().getPath();
+
+    try (Mp4Muxer muxer = new Mp4Muxer.Builder(SeekableMuxerOutput.of(outputFilePath)).build()) {
+      int firstTrackId = muxer.addTrack(FAKE_VIDEO_FORMAT);
+      int secondTrackId = muxer.addTrack(FAKE_VIDEO_FORMAT);
+      assertThrows(
+          IllegalArgumentException.class,
+          () ->
+              muxer.addTrackReference(
+                  firstTrackId, /* referenceType= */ 1234, ImmutableList.of(secondTrackId)));
+    }
   }
 
   private static void writeFakeSamples(Mp4Muxer muxer, int trackId, int sampleCount)